# samples/javascript/1-Create-ChatGPT-Plugin/public/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "OpenAPI Document on Azure Functions",
    "description": "This is the OpenAPI Document on Azure Functions",
    "version": "1.0.0"
  },
  "host": "localhost:9080",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/sqrt": {
      "get": {
        "description": "Take the square root of a number",
        "tags": [
          "Square Root"
        ],
        "operationId": "sqrt",
        "produces": [
          "application/json",
          "text/plain"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "number1",
            "description": "The number to calculate the square root of",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the square root of the number",
            "scehma": {
              "type": "string"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Error"
          }
        }
      }
    },
...
  },
...
  "securityDefinitions": {},
...
}